Minimizing main as much as possible
